pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-code-review-env"
version = "0.1.0"
description = "Production-grade OpenEnv code review benchmark with deterministic grading and realistic pull request tasks."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Rohan5commit" }]
dependencies = [
    "openenv-core[core]>=0.2.2,<1.0.0",
    "pydantic>=2.7,<3.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
server = "code_review_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["code_review_env", "code_review_env.server"]
package-dir = { "code_review_env" = ".", "code_review_env.server" = "server" }

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

